+++ lib/Scenes/Page/HomePage/Model/home_model.dart
@@ -1,1 +1,11 @@
 /// All Home Model should belongs here
+import 'package:flutter/material.dart';
+
+class TimeboxingHomePlaceholder {
+  String id;
+  bool isShow;
+  Widget widget;
+
+  TimeboxingHomePlaceholder(
+      {required this.id, required this.isShow, required this.widget});
+}
